// packages/engine-client/src/generated/ItemPlan.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Prediction } from "./Prediction";
import type { Strategy } from "./Strategy";

export type ItemPlan = { item_id: string, 
/**
 * None in Smaller mode.
 */
budget_bytes: bigint | null, strategy: Strategy, prediction: Prediction, };